[r985] Изменение : Обновление libdcadec git-v0.1.0-9-g6cc57db; Изменение : Обновление MediaInfo git-v0.7.80-5-gbb1ea54; Изменение : Обновление ZenLib git-v0.4.32-11-gf36cf81;

+++ src/ExtLib/MediaInfo/MediaInfo/Multiple/File_Lxf.cpp
@@ -104,7 +104,7 @@ const char* Lxf_PictureType[4]=
 };
 
 //---------------------------------------------------------------------------
-extern const float32 Mpegv_frame_rate[]; //In Video/File_Mpegv.cpp
+extern const float64 Mpegv_frame_rate[]; //In Video/File_Mpegv.cpp
 
 //***************************************************************************
 // Constructor/Destructor
